refactor(web): let controller own organization dialog
